Add products route
